Refactor - Remove PlayerReplicatedComponent
